add spec
